move printing functions to cheatnet
